Allow texticon to have an optional link
