+++ cmsplugin_cascade/templates/cascade/plugins/texticon.html
@@ -1,4 +1,6 @@
 {% load static sekizai_tags %}{% spaceless %}
 {% if stylesheet_url %}{% addtoblock "css" %}<link href="{{ stylesheet_url }}" rel="stylesheet" type="text/css" />{% endaddtoblock %}{% endif %}
-<i {{ instance.icon_font_class }}>{% if not instance.icon_font_class %} &square; {% endif %}</i>
+{% with instance_link=instance.link %}
+{% if instance_link %}<a href="{{ instance_link }}" {{ instance.html_tag_attributes }}>{% endif %}<i {{ instance.icon_font_class }}>{% if not instance.icon_font_class %} &square; {% endif %}</i>{% if instance_link %}</a>{% endif %}
+{% endwith %}
 {% endspaceless %}
